+++ substrate/frame/revive/rpc/examples/js/abi/errorTester.ts
@@ -1,106 +1,106 @@
 export const abi = [
-  {
-    inputs: [
-      {
-        internalType: "string",
-        name: "message",
-        type: "string",
-      },
-    ],
-    name: "CustomError",
-    type: "error",
-  },
-  {
-    inputs: [
-      {
-        internalType: "bool",
-        name: "newState",
-        type: "bool",
-      },
-    ],
-    name: "setState",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "state",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerAssertError",
-    outputs: [],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerCustomError",
-    outputs: [],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerDivisionByZero",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerOutOfBoundsError",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerRequireError",
-    outputs: [],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "triggerRevertError",
-    outputs: [],
-    stateMutability: "pure",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "value",
-        type: "uint256",
-      },
-    ],
-    name: "valueMatch",
-    outputs: [],
-    stateMutability: "payable",
-    type: "function",
-  },
-] as const;
+	{
+		inputs: [
+			{
+				internalType: 'string',
+				name: 'message',
+				type: 'string',
+			},
+		],
+		name: 'CustomError',
+		type: 'error',
+	},
+	{
+		inputs: [
+			{
+				internalType: 'bool',
+				name: 'newState',
+				type: 'bool',
+			},
+		],
+		name: 'setState',
+		outputs: [],
+		stateMutability: 'nonpayable',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'state',
+		outputs: [
+			{
+				internalType: 'bool',
+				name: '',
+				type: 'bool',
+			},
+		],
+		stateMutability: 'view',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerAssertError',
+		outputs: [],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerCustomError',
+		outputs: [],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerDivisionByZero',
+		outputs: [
+			{
+				internalType: 'uint256',
+				name: '',
+				type: 'uint256',
+			},
+		],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerOutOfBoundsError',
+		outputs: [
+			{
+				internalType: 'uint256',
+				name: '',
+				type: 'uint256',
+			},
+		],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerRequireError',
+		outputs: [],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [],
+		name: 'triggerRevertError',
+		outputs: [],
+		stateMutability: 'pure',
+		type: 'function',
+	},
+	{
+		inputs: [
+			{
+				internalType: 'uint256',
+				name: 'value',
+				type: 'uint256',
+			},
+		],
+		name: 'valueMatch',
+		outputs: [],
+		stateMutability: 'payable',
+		type: 'function',
+	},
+] as const
