+++ substrate/frame/revive/rpc/examples/js/package.json
@@ -1,22 +1,22 @@
 {
-  "name": "demo",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "ethers": "^6.13.4",
-    "prettier": "^3.3.3",
-    "solc": "^0.8.28",
-    "viem": "^2.21.47"
-  },
-  "devDependencies": {
-    "@types/bun": "^1.1.13",
-    "typescript": "^5.5.3",
-    "vite": "^5.4.8"
-  }
+	"name": "demo",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "tsc && vite build",
+		"preview": "vite preview"
+	},
+	"dependencies": {
+		"ethers": "^6.13.4",
+		"prettier": "^3.3.3",
+		"solc": "^0.8.28",
+		"viem": "^2.21.47"
+	},
+	"devDependencies": {
+		"@types/bun": "^1.1.13",
+		"typescript": "^5.5.3",
+		"vite": "^5.4.8"
+	}
 }
