+++ substrate/frame/revive/rpc/examples/js/src/balance.ts
@@ -6,4 +6,3 @@ try {
 } catch (err) {
 	console.error(err)
 }
-
